тз10

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -17,24 +17,24 @@ public class FilmController {
     private final FilmService filmService;
 
     @GetMapping("/{id}")
-    public Film getFilmById(@PathVariable Integer id) {
+    public Film getFilmById(@PathVariable Long id) {
         return filmService.getFilmById(id);
     }
 
     @PutMapping("/{id}/like/{userId}")
-    public void putLike(@PathVariable Integer id, @PathVariable Long userId) {
+    public void putLike(@PathVariable Long id, @PathVariable Long userId) {
         filmService.putLike(id, userId);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
-    public void deleteLike(@PathVariable Integer id, @PathVariable Long userId) {
+    public void deleteLike(@PathVariable Long id, @PathVariable Long userId) {
         filmService.deleteLike(id, userId);
     }
 
     @GetMapping("/popular")
-    public List<Film> getFilms(@RequestParam(required = false, name = "count") Integer count) {
+    public List<Film> getFilms(@RequestParam(required = false, name = "count") Long count) {
         if (count == null) {
-            count = 10;
+            count = 10L;
         }
         return filmService.getFilms(count);
     }
