+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -16,27 +16,27 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable Integer id) {
+    public User getUserById(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    public void addToFriendsById(@PathVariable Integer id, @PathVariable Long friendId) {
+    public void addToFriendsById(@PathVariable Long id, @PathVariable Long friendId) {
         userService.addToFriendsById(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    public void deleteFriendsById(@PathVariable Integer id, @PathVariable Long friendId) {
+    public void deleteFriendsById(@PathVariable Long id, @PathVariable Long friendId) {
         userService.deleteToFriendsById(id, friendId);
     }
 
     @GetMapping("/{id}/friends")
-    public List<User> getFriends(@PathVariable Integer id) {
+    public List<User> getFriends(@PathVariable Long id) {
         return userService.getSetFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getCommonFriends(@PathVariable Integer id, @PathVariable Long otherId) {
+    public List<User> getCommonFriends(@PathVariable Long id, @PathVariable Long otherId) {
         return userService.getCommonFriends(id, otherId);
     }
